// apps/downloader/src/modules/youtube/types/youtube-video-metadata-dto.options.ts
export type YoutubeVideoMetadataDTOOptions = {
  readonly id: string;
  readonly title: string;
  readonly thumbnail: string;
  readonly description: string;
  readonly channelId: string;
  readonly channelUrl: string;
  readonly duration: number;
  readonly viewCount: number;
  readonly webpageUrl: string;
  readonly playableInEmbed: boolean;
  readonly commentCount: number;
  readonly likeCount: number;
  readonly channel: string;
  readonly channelFollowerCount: number;
  readonly fullTitle: string;
  readonly durationString: string;
};